node_status.sh updates

+++ SebastianJ/monitoring/node_status.sh
@@ -26,11 +26,12 @@ Options:
    -t           use the Pangaea network
    -m           use the Mainnet network
    -f           disable color and text formatting
+   -z           enable debug mode
    -h           print this help
 EOT
 }
 
-while getopts "n:w:i:c:s:dtmfh" opt; do
+while getopts "n:w:i:c:s:dtmfzh" opt; do
   case ${opt} in
     n)
       node_path="${OPTARG%/}"
@@ -64,6 +65,9 @@ while getopts "n:w:i:c:s:dtmfh" opt; do
       ;;
     f)
       perform_formatting=false
+      ;;
+    z)
+      debug=true
       ;;
     h|*)
       usage
@@ -105,8 +109,11 @@ if [ -z "$maximum_block_time_difference" ]; then
 fi
 
 if [ -z "$perform_formatting" ]; then
-  # Defaults to 1 hour since last bingo if nothing else is specified
   perform_formatting=true
+fi
+
+if [ -z "$debug" ]; then
+  debug=false
 fi
 
 temp_dir="node_status"
@@ -529,7 +536,12 @@ download_file() {
   mkdir -p $temp_dir
   rm -rf "${temp_dir}/${2}"
   full_url="${1%/}/${2}"
-  wget -q $full_url --directory-prefix=$temp_dir
+  
+  if [ "$debug" = true ]; then
+    echo "Downloading ${full_url} to ${temp_dir}/${2}"
+  fi
+  
+  wget -q -O "${temp_dir}/${2}" $full_url
 }
 
 setup() {
@@ -537,7 +549,9 @@ setup() {
 }
 
 cleanup() {
-  rm -rf $temp_dir
+  if [ "$debug" = false ]; then
+    rm -rf $temp_dir
+  fi
 }
 
 #
